requirements.txt
flask
flask_cors
google-genai
groq
Pillow
requests
beautifulsoup4
python-dotenv
requests
duckduckgo-search
googletrans==3.1.0a0
gtts
wikipedia-api
speechrecognition
pyaudio
flask==2.3.3
flask-cors==4.0.0
speechrecognition==3.10.0
pyaudio==0.2.11
transformers==4.30.0
torch==2.0.1
requests==2.31.0